+++ frontend/src/containers/talent/profile/build-profile/interview-videos/myInterviewVideosScreen.js
@@ -3,13 +3,13 @@ import { Row, Col, Alert } from 'reactstrap';
 import { connect } from 'react-redux';
 import ReactPlayer from 'react-player';
 import RaisedButton from 'material-ui/RaisedButton';
-import Panel from '../../components/panel'
+import Panel from 'components/panel'
 import { withStyles } from '@material-ui/core/styles';
 import { Link } from 'react-router-dom';
 import { bindActionCreators } from 'redux';
-import * as talentActions from '../../actions/talentActions';
-import './myContactInfo.css';
-import { styles } from '../../styles';
+import * as talentActions from 'actions/talentActions';
+import '../contact-info/myContactInfo.css';
+import { styles } from 'styles';
 
 
 class MyVideos extends Component {
