+++ frontend/src/containers/talent/profile/build-profile/pictures/myPicturesScreen.css
