+++ frontend/src/containers/talent/profile/build-profile/resume/myResumeScreen.js
@@ -9,10 +9,10 @@ import ClearRounded from '@material-ui/icons/ClearRounded';
 import Dropzone from 'react-dropzone';
 import ImageLoader from 'react-loading-image';
 import ImageLightbox from 'react-image-lightbox';
-import Panel from '../../components/panel'
-import apiConfig from '../../constants/api';
-import TalentAPI from '../../apis/talentAPIs';
-import * as talentActions from '../../actions/talentActions';
+import Panel from 'components/panel'
+import apiConfig from 'constants/api';
+import TalentAPI from 'apis/talentAPIs';
+import * as talentActions from 'actions/talentActions';
 
 import 'react-image-lightbox/style.css';
 import './myResumeScreen.css';
