+++ frontend/src/containers/talent/profile/build-profile/view-profile/viewProfile.css
