+++ frontend/src/containers/talent/profile/profile-wizard/welcomeBuildProfileWizard.js
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 import RaisedButton from 'material-ui/RaisedButton';
 import Button from '@material-ui/core/Button';
 // import { withStyles } from '@material-ui/core/styles';
-import Panel from '../../../components/panel';
+import Panel from '../../../../components/panel';
 import './welcomeBuildProfileWizard.css';
